backend/requirements.txt
fastapi
uvicorn
ultralytics
torch
torchvision
python-multipart
pillow
sqlalchemy
psycopg2-binary
python-dotenv
huggingface_hub
numpy
opencv-python
redis
langchain
langchain-openai
scikit-learn
fastapi
uvicorn
ultralytics
torch
torchvision
python-multipart
pillow
sqlalchemy
psycopg2-binary
python-dotenv
huggingface_hub
numpy
opencv-python
redis
langchain
langchain-openai
scikit-learn
apscheduler
requests
opencv-python
pillow
twilio
groq
onnxruntime